Move code from Driver.swift to Execution.swift.
This addresses issues identified by SwiftLint in which Driver.swift
was deemed to long. It also allows Driver.swift to focus on
functionality given by conformance to the ParsableCommand protocol. Other
functionality not associated with command line parsing is now found
in Execution.swift.
